fix: aplicando correções do diagrama

+++ src/main/java/br/com/softhouse/dende/model/dto/response/CompraIngressoResponseDTO.java
@@ -1,10 +1,9 @@
 package br.com.softhouse.dende.model.dto.response;
 
-import br.com.softhouse.dende.model.Ingresso;
-
+import java.math.BigDecimal;
 import java.util.List;
 
 public record CompraIngressoResponseDTO(
         List<IngressoResponseDTO> ingressos,
-        double valorTotal
+        BigDecimal valorTotal
 ) {}
